text size issue
Still don't know what was the underlying cause. After further digging, the scaling issue was only attached to the home page. To patch this problem, I added a height limit so the text can't wrap underneath and is forced to stay within the desired area. Would love to know why this happened, and why it only happened on the home page